src/requirements.txt
# md-apisecurity/src/requirements.txt

aniso8601==9.0.1
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
coverage==7.1.0
exceptiongroup==1.1.0
Faker==19.6.1
Flask==2.3.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
Flask-Testing==0.8.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
packaging==23.0
pluggy==1.3.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.1
typing_extensions==4.4.0
urllib3==2.0.4
Werkzeug==2.3.7
